+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Use compatible actor IDs when generating `Like` and `Undo(Like)` activities.
 - Generate identicons for users without avatar image.
 - Generate empty banner for users without banner image.
+- Represent reply notification as `mention` with `reply` subtype.
 
 ### Fixed